Add linker script and libust initializer object
This adds a libust initializer object which gets linked automagically if
you link against libust.so, which is a linker skript from now on. The
object file just ensures that the libust is initialized properly without
forcing the user of the library to care about calling MARKER_LIB and
TRACEPOINT_LIB explicitly.

Signed-off-by: Jan Blunck <jblunck@suse.de>