Fix compilation of libust-initializer.o
The explicit rule in /Makefile.am to build libust-initializer.o was
not being considered. This is because programs that depend on it
were being built before it, triggering the use of an implicit rule
to build it.

Changing SUBDIRS in /Makefile.am to include "." insures
libust-initializer.o is built first with the right rule.