Replace __COUNTER__ with the more portable __LINE__
Also, change visibility of marker structs so they are visible
only to the current compilation unit. This allows to have the same
marker on the same line in two different files. It however does not
allow to have two identical (channel and name) markers on the same line
of the same file.