UST markers: fix structure alignment for recent gcc
GCC 4.5 (and possibly some late 4.4) choose to align structures on large
multiples, which breaks the __marker section (adding extra padding between
object sections). Increase the structure alignment to fix this issue.

Note that kernel tracepoints have recently been changed to use an array of
pointers (which are pointing to the actual tracepoint structures), thus removing
the requirement for these odd structure alignments.

But given that markers are in "maintainance mode", let's do the quick fix and
just increase the minimum alignment size.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>