Fix marker/tracepoint/trace_event lib list: expected to be sorted
Sort library lists.

List operations expect the library lists to be sorted by pointer addresses (this
was needed for iteration on kernel modules without having to hold the mutex
across read system calls). It's usefulness in userspace is debatable, but there
is clearly a bug here, since the code that iterates on the lists still expects
them to be sorted.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>