Fix up all use of /dev/stderr for portability to busybox /bin/sh
The typical shell on a small embedded target using busybox does
not have support for /dev/stderr.  A more portable way to send
output to stderr with echo is to redirect stdout to stderr with
1>&2.

In the usttrace script it did something that was effectively
a NOP.  The result of echo "" 2>/dev/stderr is not actually
going to send anything to stderr because the echo is going
to write to stdout.  This case was also fixed.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Nils Carlson <nils.carlson@ericsson.com>