Markers: temporarily remove GDB support
Remove GDB support temporarily, awaiting a better implementation that
does not require so much asm trickery. Imported the kernel LTTng code
back into UST to use C declarations for the marker structures.

The plan is to move to a SDT-based implementation (using variable
number of arguments).

This patch is required to make the MIPS port work. Tested by Jason
Wessel.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Tested-by: Jason Wessel <jason.wessel@windriver.com>