add mips support
The recent changes to UST make it trivial to add MIPS support.

The trick for MIPS is that there are many different library formats
and the gcc 4.x compiler will correctly choose the default one so long
as the libust.ldscript.in does not not make a reference to the library
format.

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>
Signed-off-by: Nils Carlson <nils.carlson@ericsson.com>