Tracepoint and TRACEPOINT_EVENT API cleanup
We want to move towards a TRACEPOINT_EVENT-based API only, so start
cleaning up right now.

Prefix with _ or __ all members that are internal to UST. Move all the
non-exported API members (that don't need to be around in program
inclusion) to tracepoint-internal.h (which is not meant to be installed
on the system).

Unit tests can still use the internal API members for now, but should
gradually move to use TRACEPOINT_EVENT as we start implementing the
FIELDS() declaration.

TRACEPOINT_EVENT is changed from the kernel incarnation (TRACE_EVENT):
we take only 3 arguments: proto, args and fields. The "fields" describe
the event layout _and_ targets the data source.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>