lttng-ust-comm: relicense to LGPLv2.1
This internal communication library is used by both lttng-sessiond and
lttng-ust. lttng-ust is LGPLv2.1, so we need to make this library
LGPLv2.1 to ensure licensing is respected.

However, we now keep this library internal to liblttng-ust and
liblttng-ust-ctl. It is not installed into the system anymore. Only the
lttng-sessiond (or another GPLv2-compatible session daemon) is meant to
control liblttng-ust through the GPLv2 liblttng-ust-ctl control library.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Signed-off-by: David Goulet <david.goulet@polymtl.ca>
Signed-off-by: Julien Desfossez <julien.desfossez@polymtl.ca>