Remove all LGPL-licensed headers
Remove all headers licensed under LGPLv2.1 for upcoming reimplementation
under the BSD-style MIT license.

Even though technically most of the content of these headers fall under
the exceptions of the LGPL:

"When a "work that uses the Library" uses material from a header file
that is part of the Library, the object code for the work may be a
derivative work of the Library even though the source code is not.
Whether this is true is especially significant if the work can be
linked without the Library, or if the work is itself a library.  The
threshold for this to be true is not precisely defined by law.

  If such an object file uses only numerical parameters, data
structure layouts and accessors, and small macros and small inline
functions (ten lines or less in length), then the use of the object
file is unrestricted, regardless of whether it is legally a derivative
work.  (Executables containing this object code plus portions of the
Library will still fall under Section 6.)"

Our intent is to make sure applications can include these headers
without having to be considered derived code.

Therefore, the license of align.h, bug.h, ringbuffer-abi.h and
ringbuffer-config.h, ust-abi.h, ust-events.h, vatomic.h are changed to
MIT license (I, Mathieu Desnoyers, am the sole author of these files and
agree to these changes).

I am removing tracepoint-event.h, tracepoint-internal.h,
tracepoint-types.h, tracepoint.h, ust-tracepoint-event-reset.h and
ust-tracepoint-event.h because those might be considered derived work
from Ftrace.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>