Split liblttng-ust into liblttng-ust and liblttng-ust-tracepoint libs
So tracepoint.h (in applications) can just dlopen
liblttng-ust-tracepoint without having to load the full liblttng-ust.

Now liblttng-ust is only needed by tracepoint probes.

This is a first step to fix the deadlock between the dynamic linker
mutex and ust mutex occurring when liblttng-ust is dlopened (due to lazy
symbol resolving of Thread-Local Storage (TLS)).

Discourage dlopen of liblttng-ust (and of tracepoint probes) in the
README.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>