Force 64-bit timestamps
commit 9080554ff25b08f48762d2ee7cfa3981317e9c1d attempted to fix the
timestamp problem by using large event headers. It incorrectly stated
that those headers "use a 64-bit timestamp field". In fact, they use a
32-bit timestamp field (not 64). So they were helping (it was better
than using a 27-bit field), but the problem could still appear. It's
only when the RING_BUFFER_RFLAG_FULL_TSC is set that the ring buffer
forces use of the 64-bit fields.

Fix this for good by setting this flag.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>