fix: on exit, leave thread/mmap reclaim to OS
Do NOT join threads: use of sys_futex makes it impossible to join the
threads without using async-cancel, but async-cancel is delivered by a
signal, which could hit the target thread anywhere in its code path,
including while the ust_lock() is held, causing a deadlock for the other
thread. Let the OS cleanup the threads if there are stalled in a
syscall.

wait_shm_mmap is used by listener threads outside of the ust lock, so we
cannot tear it down ourselves, because we cannot join on these threads.
Leave this task to the OS process exit.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>