Pack structures in comm protocol between UST and sessiond
Ensure robustness with respect to 32-bit vs 64-bit apps vs sessiond.

Since we are updating the ABI, change the order of overwrite field in
channel and channel attributes, to remove some unneeded padding.

This breaks compatibility between sessiond 2.1 and ust 2.0 (and
vice-versa), but sessiond refuses applications with version number that
does not match.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>