cyg_profile: implement fast and verbose .so
The "fast" .so (liblttng-ust-cyg-profile-fast.so) is for use-cases where
we expect a complete event stream to be recorded, so we can skip
duplicate information.

The verbose .so (liblttng-ust-cyg-profile.so) is for use-cases where
events discarded are expected, and the trace analyzer needs extra
information to be able to reconstruct the program flow.

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>