Add usage reference count for tracepoints
Keep track of how many libraries use a tracepoint, and disable the
tracepoint when the number of users drops to zero.

A new reference counter is added to tracepoint_entry. This keeps track
of how many callsites use that tracepoint.

When you have libraries and/or executables sharing tracepoints, you
cannot just disable your tracepoints when the library is unregistered.
You must check that the tracepoint is not used by any other libraries
before you disable it.

Function lib_disable_tracepoints becomes unnecessary, and is removed.

Signed-off-by: Ikaheimonen, JP <jp_ikaheimonen@mentor.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>