Fix: JUL agent connect to user and root sessiond
In order to achieve such a thing, two tracepoints are used where one is
for the regular user and the other one for root. This is to avoid
duplicating the payload in the user and root traces.

Furthermore, if a root port is found, a new thread is started to handle
the second session daemon.

This is lockstep commit with lttng-tools:
f43f95a9a82e01eed34593260d510bd32e2083ec

Fixes #732

Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>
Signed-off-by: David Goulet <dgoulet@efficios.com>