Fix linking failure when cross-compiling
Examples binaries in doc/examples require libs in liblttng-ust/.libs/,
and those libs require other libs in the same directory.  For instance
liblttng-ust.so uses liblttng-ust-tracepoint.so symbols, so the latter
is needed to properly link example programs.  Currently, ld is given
the lib dir only via -rpath (used for linking executables dependencies)
but not via -rpath-link (used for linking libs dependencies).

When compiling regularly, this does not seem to be a problem (maybe ld
is smart enough).  However when cross-compiling with the
arm-linux-gnueabihf toolchain, -rpath-link is needed otherwise linking
fails with:

arm-linux-gnueabihf/bin/ld: warning: liblttng-ust-tracepoint.so.0, needed by ../../../liblttng-ust/.libs//liblttng-ust.so, not found (try using -rpath or -rpath-link)

This patch adds -rpath-link to doc/examples/Makefile.am.

Signed-off-by: Adrien Vergé <adrienverge@gmail.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>